GatedResourceLayout, Recording version
